Adding an E2E test for the member access condition
