+++ packages/contracts/test/unit-testing/common.ts
@@ -6,6 +6,8 @@ export const abiCoder = ethers.utils.defaultAbiCoder;
 export const EMPTY_DATA = '0x';
 export const ZERO_BYTES32 =
   '0x0000000000000000000000000000000000000000000000000000000000000000';
+export const ONE_BYTES32 =
+  '0x0000000000000000000000000000000000000000000000000000000000000001';
 
 export const DEPLOYER_PERMISSION_ID = ethers.utils.id('DEPLOYER_PERMISSION');
 export const EDITOR_PERMISSION_ID = ethers.utils.id('EDITOR_PERMISSION');
